Enable reading of UNKNOWN solution
